lee 修改优惠券

+++ MySuperApp/Class/Databsae/other/EADataModel/优惠券列表页/CouponcardListCouponcardListModel.h
@@ -18,7 +18,7 @@
 @property (nonatomic, retain) NSArray *checkoutCards;
 @property (nonatomic, retain) NSArray *checkoutCouponcard;
 @property (nonatomic, assign) int cards_count;//卡个数
-@property (nonatomic, assign) int couponcard_count;//优惠券个数
+@property (nonatomic, assign) NSInteger couponcard_count;//优惠券个数
 
 
 @property (nonatomic, assign) int requestTag;
